# apps/api-lab/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Library Book Management API
  version: 1.0.0
  description: |
    API for managing library books and reservations.

    ## Versioning Strategy

    This API uses URI-based versioning (e.g., `/api/v1/...`). When breaking changes are
    introduced, a new version prefix will be added (e.g., `/api/v2/...`). Previous versions
    will enter a sunset period of at least 6 months, during which they remain functional but
    return a `Sunset` header indicating the deprecation date. Clients should monitor this
    header and migrate promptly.

    ## Pagination

    List endpoints use cursor-based pagination via `limit` and `continuation_token` query
    parameters. The response includes a `continuation_token` field (opaque string) and a
    `has_more` boolean. To fetch the next page, pass the returned `continuation_token` as a
    query parameter.

    ## Idempotency

    POST endpoints accept an `Idempotency-Key` header (UUID). If a request with the same
    idempotency key has already been processed, the server returns the original response
    without creating a duplicate resource. Keys expire after 24 hours.

    ## Rate Limiting

    All endpoints are rate-limited. Every response includes `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers. When the limit is exceeded,
    the server returns 429 Too Many Requests with a `Retry-After` header.
  contact:
    name: API Lab Team
    email: dev@null.com
  license:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      security: []
      tags:
        - auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 100
                  pattern: "^[a-zA-Z0-9_-]+$"
                email:
                  type: string
                  minLength: 5
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  username:
                    type: string
        "409":
          description: Username already exists
  /api/v1/auth/login:
    post:
      security: []
      tags:
        - auth
      summary: Authenticate and obtain tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                  token_type:
                    type: string
        "401":
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      security: []
      tags:
        - auth
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/paths/~1api~1v1~1auth~1login/post/responses/200/content/application~1json/schema"
        "401":
          description: Invalid refresh token
  /api/v1/auth/logout:
    post:
      security: []
      tags:
        - auth
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        "204":
          description: Logged out
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the auth service login endpoint
...